chore: add core cost and itinerary modules
